toolkit/components/passwordmgr/test/mochitest/test_autofill_from_bfcache.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test autofilling documents restored from the back/forward cache (bfcache)</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="pwmgr_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<script>
runChecksAfterCommonInit();

runInParent(function initLogins() {
  const {Services} = ChromeUtils.import("resource://gre/modules/Services.jsm");

  let login1 = Cc["@mozilla.org/login-manager/loginInfo;1"].createInstance(Ci.nsILoginInfo);
  login1.init("https://example.com", "", null, "autofilled", "pass1", "", "");
  Services.logins.addLogin(login1);
});

</script>
<p id="display"></p>

<div id="content">
  <a id="next" href="https://example.org/" target="loginWin">Next</a>
</div>

<pre id="test">
<script class="testbody" type="text/javascript">
let win;

add_task(async function test_crossOriginBfcacheRestore() {
  let processedPromise = promiseFormsProcessed();
  win = window.open("form_basic.html", "loginWin");
  await processedPromise;
  let doc = win.document;
  let uname = doc.getElementById("form-basic-username");
  let pword = doc.getElementById("form-basic-password");
  checkLoginForm(uname, "autofilled", pword, "pass1");

  let pageHidePromise = new Promise(resolve => {
    win.addEventListener("pagehide", resolve, {once: true});
  });
  info("Navigate to another page on a different origin");
  document.getElementById("next").click();
  let pageHideEvent = await pageHidePromise;
  await SimpleTest.promiseFocus(win);
  ok(pageHideEvent.persisted, "First document should have been persisted in bfcache");

  info("go back to the first page (login form)");
  processedPromise = promiseFormsProcessed();
  SpecialPowers.wrap(win).history.back();
  await processedPromise;
  checkLoginForm(uname, "autofilled", pword, "pass1");
  win.close();
});
</script>
</pre>
</body>
</html>
